adding delete features

+++ src/screens/AddRemedy/index.js
@@ -1,53 +1,59 @@
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import { useNavigation } from '@react-navigation/native';
-import React, { useState } from 'react';
-import { ScrollView, View, Text, Image } from 'react-native';
-import { TextInput, Button } from 'react-native-paper';
-import uuid from 'react-native-uuid';
-import { v4 as uuidv4 } from 'uuid';
-import Pills from '../../assets/icons/medicine.png';
-import Watch from '../../assets/icons/watch.png';
-import { Header } from '../../components/Header';
-import 'react-native-get-random-values';
-import { COLORS } from '../../theme';
-import { globalStyles } from '../../theme/globalStyles';
-import { styles } from './styles';
+import AsyncStorage from "@react-native-async-storage/async-storage";
+import { useNavigation } from "@react-navigation/native";
+import React, { useState } from "react";
+import { ScrollView, View, Text } from "react-native";
+import { TextInput, Button } from "react-native-paper";
+import { v4 as uuidv4 } from "uuid";
+import { Header } from "../../components/Header";
+import "react-native-get-random-values";
+import { COLORS } from "../../theme";
+import { globalStyles } from "../../theme/globalStyles";
+import { styles } from "./styles";
+import useSnackBar from "../../hooks/useSnackbar";
 
 export function AddRemedy() {
   const navigation = useNavigation();
   const [name, setName] = useState();
   const [loading, setLoading] = useState(false);
+  const { addSnackbar } = useSnackBar();
 
+  // eslint-disable-next-line consistent-return
   const getList = async () => {
     try {
-      const list = await AsyncStorage.getItem('PILLS');
+      const list = await AsyncStorage.getItem("PILLS");
       if (list) {
         return JSON.parse(list);
-      } else {
-        return [];
       }
+      return [];
     } catch (err) {
       console.debug(err);
     }
   };
 
-  const saveRemedy = async (name) => {
+  const saveRemedy = async (remedyName) => {
     setLoading(true);
     const newRemedy = {
-      name,
+      name: remedyName,
       id: uuidv4(),
+      hours: [],
     };
     try {
-      console.debug('AQUI');
       const existingList = await getList();
-      console.debug(existingList);
+
+      if (existingList.length === 6) {
+        setLoading(false);
+        addSnackbar(`Já existe um limite máximo de 6 remédios cadastrados`);
+        return;
+      }
 
       const updatedList = [...existingList, newRemedy]; // notice the newData here
-      console.debug(updatedList);
-      await AsyncStorage.setItem('PILLS', JSON.stringify(updatedList));
+      await AsyncStorage.setItem("PILLS", JSON.stringify(updatedList));
       setLoading(false);
-      navigation.navigate('Home');
-    } catch (err) {}
+      addSnackbar(`${remedyName} adicionado com sucesso`);
+      navigation.navigate("Home");
+    } catch (err) {
+      addSnackbar(`Erro ao adicionar remédio`);
+    }
   };
 
   return (
@@ -55,12 +61,14 @@ export function AddRemedy() {
       <Header />
       <ScrollView>
         <View style={styles.container}>
-          <Text style={[globalStyles.title, styles.title]}>Adicionar Remédio</Text>
+          <Text style={[globalStyles.title, styles.title]}>
+            Adicionar Remédio
+          </Text>
           <TextInput
             label="Nome do Remédio"
             value={name}
             activeUnderlineColor={COLORS.LIGHT_BLUE}
-            onChangeText={(name) => setName(name)}
+            onChangeText={(text) => setName(text)}
           />
           <Button
             icon="content-save"
@@ -68,7 +76,8 @@ export function AddRemedy() {
             style={{ marginTop: 30 }}
             mode="contained"
             loading={loading}
-            onPress={() => saveRemedy(name)}>
+            onPress={() => saveRemedy(name)}
+          >
             Adicionar
           </Button>
         </View>
